Feature/projectbackend

+++ backend/src/main/java/edu/cit/projectsync/Controller/TaskController.java
@@ -30,7 +30,7 @@ public class TaskController {
     private ProjectService projectService;
 
     @PostMapping("/createtask")
-    public ResponseEntity<TaskEntity> createTask(@RequestBody TaskEntity task, @RequestParam Long projectId) {
+    public ResponseEntity<TaskEntity> createTask(@RequestBody TaskEntity task, @RequestParam int projectId) {
         ProjectEntity project = projectService.getProjectById(projectId);
         if (project == null) {
             return ResponseEntity.badRequest().build(); // Return 400 if the project doesn't exist
@@ -43,7 +43,7 @@ public class TaskController {
     }
 
     @PutMapping("/updatetask/{taskId}/")
-    public ResponseEntity<TaskEntity> updateTask(@PathVariable Long taskId, @RequestBody TaskEntity updatedTask) {
+    public ResponseEntity<TaskEntity> updateTask(@PathVariable int taskId, @RequestBody TaskEntity updatedTask) {
         TaskEntity task = taskService.updateTask(taskId, updatedTask);
         if (task != null) {
             return ResponseEntity.ok(task);
@@ -52,7 +52,7 @@ public class TaskController {
     }
 
     @GetMapping("/gettaskid/{taskId}/")
-    public ResponseEntity<TaskEntity> getTaskById(@PathVariable Long taskId) {
+    public ResponseEntity<TaskEntity> getTaskById(@PathVariable int taskId) {
         TaskEntity task = taskService.getTaskById(taskId);
         if (task != null) {
             return ResponseEntity.ok(task);
@@ -61,7 +61,7 @@ public class TaskController {
     }
 
     @GetMapping("/project/{projectId}/")
-    public ResponseEntity<List<TaskEntity>> getTasksByProjectId(@PathVariable Long projectId) {
+    public ResponseEntity<List<TaskEntity>> getTasksByProjectId(@PathVariable int projectId) {
         List<TaskEntity> tasks = taskService.getTasksByProjectId(projectId);
         if (tasks != null && !tasks.isEmpty()) {
             return ResponseEntity.ok(tasks);
@@ -76,7 +76,7 @@ public class TaskController {
     }
 
     @DeleteMapping("/deletetask/{taskId}/")
-    public ResponseEntity<Void> deleteTask(@PathVariable Long taskId) {
+    public ResponseEntity<Void> deleteTask(@PathVariable int taskId) {
         taskService.deleteTask(taskId);
         return ResponseEntity.noContent().build();
     }
